requirements.txt
fastapi>=0.109.0
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
httpx>=0.25.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
anyio>=4.6.0
click>=8.1.0
starlette>=0.27.0
mcp[cli]>=0.1.0 